move: parseFromInto test to a separate class
