+++ src/test/java/module-info.java
@@ -6,4 +6,5 @@ module lanat.test {
 	exports lanat.test to org.junit.platform.commons, lanat;
 	exports lanat.test.manualTests to org.junit.platform.commons, lanat;
 	exports lanat.test.units to lanat, org.junit.platform.commons;
+	exports lanat.test.units.commandTemplates to lanat, org.junit.platform.commons;
 }
